pyproject.toml
[build-system]
requires = ["flit_core >=3.10,<4"]
build-backend = "flit_core.buildapi"

[project]
keywords = ['python']
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",

    "License :: OSI Approved :: MIT License",

]
authors = [{ name = "Chad Lowe", email = "pfmsoft.dev@gmail.com" }]
name = "pbs-split"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
dynamic = ["version", "description"]
# https://chrisholdgraf.com/blog/2022/install-github-from-pyproject/
# https://peps.python.org/pep-0508/#examples
# "pfmsoft-snippets@git+https://github.com/donalchilde/pfmsoft-snippets@0.0.1"
dependencies = [
    "typer",
    "pfmsoft-snippets-indexed-string@git+https://github.com/donalchilde/pfmsoft-snippets-indexed-string@dev",
    "pfmsoft-snippets-simple-serializer@git+https://github.com/donalchilde/pfmsoft-snippets-simple-serializer@dev",
]

[project.scripts]
pbs-split = "pbs_split.cli.main_typer:app"

[project.urls]
Documentation = "https://github.com/DonalChilde/pbs-split#readme"
# Documentation_rtd = "https://pbs-split.readthedocs.io"
Issues = "https://github.com/DonalChilde/pbs-split/issues"
Source = "https://github.com/DonalChilde/pbs-split"


[project.optional-dependencies]
dev = [

    # Automation
    "nox",
    "pre-commit",

    # Linting and Formatting
    "black",
    "isort",

    # Tools
    "pysnooper",
    "scriv",
]
lint = ["pylint", "mypy"]
doc = [
    # Documentation
    "Sphinx",
    "sphinx-autobuild",
    "sphinx-autodoc-typehints",
    "sphinx_rtd_theme",
    "myst-parser",
    "sphinxcontrib-typer",
]
vscode = ["esbonio", "rst2html", "rstcheck"]
testing = ["pytest", "coverage[toml]", "pytest-cov"]


[tool.isort]
profile = "black"
src_paths = ["src", "test"]

[tool.black]
line-length = 88
target-version = ['py313']

# https://pylint.pycqa.org/en/latest/user_guide/usage/run.html#command-line-options
# https://pylint.pycqa.org/en/latest/user_guide/configuration/all-options.html#standard-checkers
# pylint --generate-toml-config > ./pylint-toml-example.toml
# [tool.pylint.messages_control]
# disable = [
#     "raw-checker-failed",
#     "bad-inline-option",
#     "locally-disabled",
#     "file-ignored",
#     "suppressed-message",
#     "useless-suppression",
#     "deprecated-pragma",
#     "use-symbolic-message-instead"
# ]

[tool.pylint.format]
# for black compatability
max-line-length = "95"

[tool.pytest.ini_options]
pythonpath = "src"

[tool.coverage.run]
branch = true
parallel = true
# omit = [""]

[tool.scriv]
#version = "literal: pyproject.toml: project.version"
version = "literal: src/pbs_split/__init__.py: __version__"
format = "md"
md_header_level = "2"
entry_title_template = "file: changelog.d/templates/entry_title_template.md.j2"
new_fragment_template = "file: changelog.d/templates/new_fragment.${config:format}.j2"
